% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopulationFractions.R
\name{getPopulationFractions}
\alias{getPopulationFractions}
\title{Get population fractions}
\usage{
getPopulationFractions(ages, year = 2015, population)
}
\arguments{
\item{ages}{Vector of ages. Each element represents the upper range of an age range. The lowest bound is presumed to be zero.
If the oldest age does not reach the end of the population range, an additional element is added to span the full range.
This function by default uses \code{flumodelsutil_data$population.US} for age information (2015 US population).
One could also use the \code{acs} package to download data dynamically, but that requires the user to request an individual
key from the ACS for access.}

\item{year}{Year to sample population from. Defaults to 2015.}

\item{population}{Population data frame that contains age and population values. Defaults to 2015 US population.
The two required columns of the data frame are AGE and TOT_POP.}
}
\value{
A vector population fractions that sums to 1
}
\description{
Gets population fractions for a given set of age ranges
}
\author{
Matt Clay <clay.matt@gmail.com>
}
