% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVaccineUptakeMultiplierFromAllocation.R
\name{getVaccineUptakeMultiplierFromAllocation}
\alias{getVaccineUptakeMultiplierFromAllocation}
\title{Get vaccine uptake multiplier from overall vaccine allocation fractions}
\usage{
getVaccineUptakeMultiplierFromAllocation(
  ages,
  population = flumodelsutil::flumodelsutil_data$population.US \%>\% dplyr::filter(YEAR
    == year & MONTH == 1 & AGE != 999) \%>\% dplyr::select(AGE, TOT_POP),
  year = 2015,
  originalAllocation,
  originalAllocationAges
)
}
\arguments{
\item{ages}{Vector of ages. Each element represents the upper range of an age range. The lowest bound is presumed to be zero.
If the oldest age does not reach the end of the population range, an additional element is added to span the full range.
The final age bracket cannot start after the final bracket of originalContactMatrixAges.}

\item{population}{Population data frame that contains age and population values. Defaults to 2015 US population.
The two required columns of the data frame are AGE and TOT_POP}

\item{year}{Year to sample population from. Defaults to 2015.}

\item{originalAllocation}{Original allocation fraction of vaccine, by age group.}

\item{originalAllocationAges}{Ages associated with originalAllocation. Must be in same format as ages.}
}
\value{
Vector of vaccine uptake multipliers.
}
\description{
Calculates vaccine uptake multipliers suitable for use in flumodels::SEIRVModel. It takes
   vaccine allocation by age group (e.g., 30% to <18, 40% to 19-60, and 30% to 61+) and converts that into
   relative uptakes for arbitrary age bins.
}
\author{
Matt Clay <clay.matt@gmail.com>
}
